add upload function

+++ app/Http/Controllers/CategoryController.php
@@ -48,9 +48,9 @@ class CategoryController extends Controller
     }
 
     public function destroy($id){
-        $category = Category::withCount(['child'])->find($id);
+        $category = Category::withCount(['child', 'product'])->find($id);
 
-        if ($category->child_count == 0) {
+        if ($category->child_count == 0 && $category->product_count == 0) {
             $category->delete();
             return redirect(route('category.index'))->with(['succes' => 'Kategori Dihapus!']);
         }
